♻️ [refactor]: #452 Phase A+B: lexer_read_array_dict.rs を分離
read_array_begin/end の 11 テスト + read_dict_begin/end の 14 テスト
（合計 25 件）を `src/lexer/tests/lexer_read_array_dict.rs` に分離。
テスト本体で `Token::ArrayBegin` 等を直接参照するため
`use super::super::token::Token;` を追加。

- 293 件グリーン維持
- 短縮テスト関数名集合不変 / 重複 0
- lexer_read_array_dict.rs: 216 行 (< 800)

Refs #452
